requirements.txt
spacy==3.7.2
pdfplumber==0.11.0
streamlit==1.29.0
scikit-learn==1.4.0
PyYAML==6.0.1
joblib==1.3.2
pytest==7.4.3